+++ packages/thumbprint-react/components/Avatar/index.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
+import isNumber from 'lodash/isNumber';
 import * as tokens from '@thumbtack/thumbprint-tokens';
 import Badge from './subcomponents/badge.jsx';
 import styles from './index.module.scss';
@@ -14,14 +15,12 @@ const CheckIcon = () => (
     </svg>
 );
 
-const getBadgeProps = ({ size, hasUnreadNotifications, isChecked, isOnline }) => {
+const getBadgeProps = ({ size, isChecked, isOnline }) => {
     const props = {
         size,
     };
 
-    if (hasUnreadNotifications) {
-        props.background = 'red';
-    } else if (isChecked) {
+    if (isChecked) {
         props.children = <CheckIcon />;
         props.background = 'green';
     } else if (isOnline) {
@@ -31,8 +30,8 @@ const getBadgeProps = ({ size, hasUnreadNotifications, isChecked, isOnline }) =>
     return props;
 };
 
-const shouldShowBadge = ({ size, hasUnreadNotifications, isChecked, isOnline }) =>
-    size !== 'xsmall' && (hasUnreadNotifications || isChecked || isOnline);
+const shouldShowBadge = ({ size, isChecked, isOnline }) =>
+    size !== 'xsmall' && (isChecked || isOnline);
 
 const STYLES = [
     {
@@ -90,11 +89,13 @@ class EntityAvatar extends React.Component {
         return (
             <div
                 className={classNames(styles.root, {
+                    [styles.rootXsmall]: size === 'xsmall',
                     [styles.rootSmall]: size === 'small',
                     [styles.rootMedium]: size === 'medium',
                     [styles.rootLarge]: size === 'large',
                     [styles.rootXlarge]: size === 'xlarge',
                 })}
+                style={isNumber(size) ? { width: size, height: size } : {}}
             >
                 {imageUrl ? (
                     <img
@@ -134,14 +135,17 @@ EntityAvatar.propTypes = {
     /**
      * The set of avatar sizes that we support.
      */
-    size: PropTypes.oneOf(['small', 'medium', 'large', 'xlarge']),
+    size: PropTypes.oneOfType([
+        PropTypes.oneOf(['xsmall', 'small', 'medium', 'large', 'xlarge']),
+        PropTypes.number,
+    ]),
 };
 
 EntityAvatar.defaultProps = {
     imageUrl: undefined,
     initial: undefined,
     fullName: undefined,
-    size: 'medium',
+    size: 'small',
 };
 
 // TODO(giles): remove this default export once website has been updated to refer only to UserAvatar
@@ -165,6 +169,7 @@ export default class UserAvatar extends React.Component {
 
     render() {
         const { props } = this;
+
         return (
             <div
                 className={classNames(styles.root, {
@@ -174,6 +179,7 @@ export default class UserAvatar extends React.Component {
                     [styles.rootLarge]: props.size === 'large',
                     [styles.rootXlarge]: props.size === 'xlarge',
                 })}
+                style={isNumber(props.size) ? { width: props.size, height: props.size } : {}}
             >
                 {shouldShowBadge(props) && <Badge {...getBadgeProps(props)} />}
                 {props.imageUrl ? (
@@ -215,16 +221,14 @@ UserAvatar.propTypes = {
     /**
      * The set of avatar sizes that we support.
      */
-    size: PropTypes.oneOf(['xsmall', 'small', 'medium', 'large', 'xlarge']),
+    size: PropTypes.oneOfType([
+        PropTypes.oneOf(['small', 'medium', 'large', 'xlarge']),
+        PropTypes.number,
+    ]),
     /**
      * Displays a badge of a checkmark next to the `Avatar`.
      */
     isChecked: PropTypes.bool,
-    /**
-     * Should be true if the user has unread notifications. Displays a badge
-     * as an indication.
-     */
-    hasUnreadNotifications: PropTypes.bool,
     /**
      * Displays a badge if the user is online.
      */
@@ -235,9 +239,8 @@ UserAvatar.defaultProps = {
     imageUrl: undefined,
     initials: undefined,
     fullName: undefined,
-    size: 'medium',
+    size: 'small',
     isChecked: false,
-    hasUnreadNotifications: false,
     isOnline: undefined,
 };