+++ packages/thumbprint-codemods/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 0.3.2 - 2019-06-12
+
+### Fixed
+
+-   [Patch] Exit instead of returning in the CLI.
+
 ## 0.3.1 - 2019-06-11
 
 ### Fixed